# backend/src/api-gateway/routes/curriculum/institutions.routes.ts
// AUTO-GENERATED by pzo_curriculum_gateway_generator.py — DO NOT EDIT MANUALLY
import { Router, Request, Response, NextFunction } from 'express';
import { requireAuth, requireScope, requireInstitutionAccess } from '../../middleware/auth';
import { validateBody } from '../../middleware/validation';
import { rateLimit } from '../../middleware/security';
import { ok, created, fail, AppError } from '../../middleware/errors';
import { emitAudit, CurriculumAuditEvent } from '../../middleware/audit';
import { licensingControlPlaneAdapter } from '../../adapters/licensing_control_plane.adapter';
import { CreateInstitutionSchema, UpdateInstitutionSchema } from '../../contracts/curriculum/institution.dto';

const router = Router();

router.post('/',
  requireAuth,
  requireScope('curriculum:institutions:write'),
  rateLimit('write'),
  validateBody('CreateInstitution', CreateInstitutionSchema),
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const ctx = req.tenantContext!;
      const institution = await licensingControlPlaneAdapter.createInstitution(ctx, req.body);
      await emitAudit(req, CurriculumAuditEvent.INSTITUTION_CREATED, 'institution', institution.id, { name: institution.name });
      res.status(201).json(created(req, institution));
    } catch (err) {
      next(err);
    }
  },
);

router.get('/:institutionId',
  requireAuth,
  requireInstitutionAccess,
  rateLimit('read'),
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const ctx = req.tenantContext!;
      const institution = await licensingControlPlaneAdapter.getInstitution(ctx, req.params.institutionId);
      if (!institution) {
        res.status(404).json(fail(req, 'NOT_FOUND', 'Institution not found'));
        return;
      }
      res.json(ok(req, institution));
    } catch (err) {
      next(err);
    }
  },
);

router.patch('/:institutionId',
  requireAuth,
  requireInstitutionAccess,
  requireScope('curriculum:institutions:write'),
  rateLimit('write'),
  validateBody('UpdateInstitution', UpdateInstitutionSchema),
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const ctx = req.tenantContext!;
      const institution = await licensingControlPlaneAdapter.updateInstitution(ctx, req.params.institutionId, req.body);
      if (!institution) {
        res.status(404).json(fail(req, 'NOT_FOUND', 'Institution not found'));
        return;
      }
      await emitAudit(req, CurriculumAuditEvent.INSTITUTION_UPDATED, 'institution', institution.id);
      res.json(ok(req, institution));
    } catch (err) {
      next(err);
    }
  },
);

router.get('/',
  requireAuth,
  requireScope('curriculum:institutions:read'),
  rateLimit('read'),
  async (req: Request, res: Response, next: NextFunction) => {
    try {
      const ctx = req.tenantContext!;
      const result = await licensingControlPlaneAdapter.listInstitutions(ctx, {
        page: parseInt(req.query.page as string || '1', 10),
        pageSize: parseInt(req.query.pageSize as string || '20', 10),
      });
      res.json(ok(req, result.items, result.meta));
    } catch (err) {
      next(err);
    }
  },
);

export { router as institutionsRouter };