# openapi/curriculum.v1.yaml
openapi: 3.1.0
info:
  title: Point Zero One — Curriculum Gateway API
  version: 1.0.0
  description: |
    B2B curriculum gateway for institutional delivery of Point Zero One financial
    literacy simulation. Supports cohort management, benchmark administration,
    measurement/reporting, and facilitator tooling.
  contact:
    name: Density6 LLC
    url: https://pointzeroonegame.com
servers:
  - url: https://pointzeroonegame.com/api/v1/curriculum
    description: Production
  - url: http://localhost:3000/api/v1/curriculum
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
  /institutions:
    post:
      summary: Create institution
      tags: [Institutions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name:
                  type: string
                slug:
                  type: string
      responses:
        '201':
          description: Institution created
        '400':
          description: Validation error
        '401':
          description: Unauthorized
    get:
      summary: List institutions
      tags: [Institutions]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of institutions
...